+++ src/ECommerce.Domain/Entities/CartItem.cs
@@ -7,7 +7,7 @@ public class CartItem : BaseEntity
     public Guid CartId { get; set; }
     public Guid ProductId { get; set; }
     public int Quantity { get; set; }
-    public double UnitPriceSnapshot { get; set; }
+    public decimal UnitPriceSnapshot { get; set; }
 
     public Cart Cart { get; set; } = default!;
     public Product Product { get; set; } = default!;
@@ -16,7 +16,7 @@ public class CartItem : BaseEntity
 
     public static CartItem CreateStub(Guid id) => new() { Id = id };
 
-    public static CartItem Create(Guid cartId, Guid productId, int quantity, double unitPriceSnapshot) =>
+    public static CartItem Create(Guid cartId, Guid productId, int quantity, decimal unitPriceSnapshot) =>
         new()
         {
             CartId = cartId,
